+++ levelEditor/control.js
@@ -485,8 +485,16 @@ document.addEventListener("keydown", function (input) {
                   level.length * baseBlockSize,
                   window.innerWidth + 2 * camOffsetLimit
                 );
+                id("bgLayer").width = Math.min(
+                  level.length * baseBlockSize,
+                  window.innerWidth + 2 * camOffsetLimit
+                );
                 id("lvlHeight").innerHTML = level[0].length;
                 id("levelLayer").height = Math.min(
+                  level[0].length * baseBlockSize,
+                  window.innerHeight + 2 * camOffsetLimit
+                );
+                id("bgLayer").height = Math.min(
                   level[0].length * baseBlockSize,
                   window.innerHeight + 2 * camOffsetLimit
                 );
@@ -506,8 +514,16 @@ document.addEventListener("keydown", function (input) {
                 level.length * baseBlockSize,
                 window.innerWidth + 2 * camOffsetLimit
               );
+              id("bgLayer").width = Math.min(
+                level.length * baseBlockSize,
+                window.innerWidth + 2 * camOffsetLimit
+              );
               id("lvlHeight").innerHTML = level[0].length;
               id("levelLayer").height = Math.min(
+                level[0].length * baseBlockSize,
+                window.innerHeight + 2 * camOffsetLimit
+              );
+              id("bgLayer").height = Math.min(
                 level[0].length * baseBlockSize,
                 window.innerHeight + 2 * camOffsetLimit
               );